New disc-bounding-box intersector; not integrated into tracer yet.
Note: This version will compute the intersection of a disc with only one
single bounding plain. If the disc intersects more than one bounding
plain, the code will not deliver correct results or an assertion error.
Extending the code to handle multiple plains intersecting the disc
shouldn't be difficult; see also code snipets at the bottom of the
intersector class.